split module further into tool, variable and want submodules
